+++ node/CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Remove unsafe consensus spin-time environment mutation during node initialization
+- Reject deploy transactions whose gas charge overflows during mempool
+  admission.
+- Reject blob transactions whose gas charge overflows during mempool
+  admission
 
 ## [1.6.0] - 2026-02-27